requirements.txt
fastapi
uvicorn
requests
python-dotenv
azure-functions
